Fixed xpath error when running under trac 0.10 .
git-svn-id: https://trac-hacks.org/svn@7910 7322e99d-02ea-0310-aa39-e9a107903beb

+++ hudsontracplugin/0.10/HudsonTrac/HudsonTracPlugin.py
@@ -144,7 +144,7 @@ class HudsonTracPlugin(Component):
         add_stylesheet(req, 'HudsonTrac/hudsontrac.css')
 
         # get and parse the build-info
-        url = self.info_url % {'start': start*1000, 'stop': stop*1000}
+        url = self.info_url % {'start': int(start*1000), 'stop': int(stop*1000)}
         try:
             try:
                 info = minidom.parse(self.url_opener.open(url))
